requirements.txt
Flask
pdfplumber
python-docx
scikit-learn
gunicorn
requests
python-dotenv
reportlab
firebase-admin
python-dotenv